requirements.txt
python-dotenv
beautifulsoup4
requests
SQLAlchemy
psycopg2-binary
etherscan-python
Flask
